Added helpers, MessagePasser infrastructure, created test project
Added some helper classes and methods that are needed or will be needed in the
project.

Created infrastructure for MessagePasser in the Communication namespace. This
is yet to be tested and refined.

Added tests project that will be used for unit testing all of our classes.

Added TODO file noting the short-term tasks.

Signed-off-by: Filip <vilaemail@gmail.com>