Initial implementation of Logging mechanism.
Added Logging subsystem with logging functionality. Wrote some of the unit
tests for the subsystem.
Changed Program.cs to instantiate a logger and dispose on app exit.
Now we are catching unhandled exceptions and logging the exception before exit.

Still need to:
Write all tests for logger subsystem.
Document the new code.

Signed-off-by: Filip <vilaemail@gmail.com>